Support configuration from environment variables (#2)
* Support loading Config from environment variables

This uses "envy" for convenience and reuse of the existing serde-enabled
structs. File-based configuration still takes priority, and envy is only
called if the `movine.toml` file cannot be loaded. Any errors when
loading from the environment are ignored, and the file loading error
will be returned.

* Add dotenv support to CLI

This is a popular mechanism for configuring Rust CLIs, and simplifies
using environment-variable based configuration.